Update the moa_view_types module
Use an alternative method to construct the views to avoid self-referencing. This method may fail with gfortran as that compiler seems to implement default assignment via recursive data types incorrectly (at least this was the case in a small test program).
The associated batch file and the test program can be used for a demonstration